[core] Fix issues with sql->Async

See merge request wings-admin/wings-lsb!27

+++ src/map/entities/battleentity.cpp
@@ -19,8 +19,8 @@
 ===========================================================================
 */
 
-#include "../../common/logging.h"
-#include "../../common/utils.h"
+#include "common/logging.h"
+#include "common/utils.h"
 
 #include "battleentity.h"
 
@@ -1750,7 +1750,7 @@ bool CBattleEntity::OnAttack(CAttackState& state, action_t& action)
                         if (PTarget->objtype == TYPE_PC)
                         {
                             auto* targ_weapon = dynamic_cast<CItemWeapon*>(PTarget->m_Weapons[SLOT_MAIN]);
-                            uint8 skilltype   = (targ_weapon == nullptr ? SKILL_HAND_TO_HAND : targ_weapon->getSkillType());
+                            uint8 skilltype   = (targ_weapon == nullptr ? (uint8)SKILL_HAND_TO_HAND : targ_weapon->getSkillType());
                             charutils::TrySkillUP((CCharEntity*)PTarget, (SKILLTYPE)skilltype, GetMLevel());
                         } // In case the Automaton can counter
                         else if (PTarget->objtype == TYPE_PET && PTarget->PMaster && PTarget->PMaster->objtype == TYPE_PC &&
